+++ routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use Azuriom\Plugin\Sitemap\Controllers\SitemapController;
+use Azuriom\Plugin\Seo\Controllers\SitemapController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/sitemap.xml', [SitemapController::class, 'index'])->name('index');